task: Start refactoring shell builder
